requirements.txt
streamlit
pytesseract
pillow
python-docx
